+++ users/views/user_diag_record_views.py
@@ -1,11 +1,13 @@
 from django.db import transaction
 from drf_yasg.utils import swagger_auto_schema
+from rest_framework import status
 from rest_framework.views import APIView
 
 from config.basemodel import ApiResponse
 from config.basemodel import validator
 from config.settings import REQUEST_BODY, REQUEST_QUERY
-from users.serializers import *
+from users.serializers.user_diag_serializers import *
+from users.serializers.user_get_post_put_delete_serializers import UserIdReqeust
 
 
 class RecordSaveView(APIView):
