+++ users/views/user_sign_up_views.py
@@ -1,11 +1,12 @@
 from django.db import transaction
 from drf_yasg.utils import swagger_auto_schema
+from rest_framework import status
 from rest_framework.views import APIView
 
 from config.basemodel import ApiResponse
 from config.basemodel import validator
 from config.settings import REQUEST_BODY
-from users.serializers import *
+from users.serializers.user_get_post_put_delete_serializers import *
 
 
 class SignupView(APIView):
